# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: QProv API
  contact:
    name: GitHub Repository
    url: https://github.com/UST-QuAntiL/QProv
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.3.0
servers:
- url: http://localhost:5020/qprov
  description: Generated server url
paths:
...
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}:
    get:
      tags:
      - provider
      description: Retrieve a specific Qubit and its basic properties.
      operationId: getQubit
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQubitDto'
        "400":
          description: Qubit belongs not to specified provider and QPU.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQubitDto'
        "404":
          description: Not Found. Qubit with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelQubitDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates:
    get:
      tags:
      - provider
      description: Retrieve all Gates that can be executed on the Qubit.
      operationId: getGates
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateDto'
        "404":
          description: Qubit or QPU not available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates/{gateId}:
    get:
      tags:
      - provider
      description: Retrieve a specific Qubit and its basic properties.
      operationId: getGate
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: gateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelGateDto'
        "404":
          description: Not Found. Gate with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelGateDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates/{gateId}/characteristics:
    get:
      tags:
      - provider
      description: "Retrieve the calibration characteristics from the given gate.\
        \ By using the latest parameter only the latest data is retrieved, otherwise\
        \ all available data."
      operationId: getGateCharacterisitcs
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: gateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateCharacteristicsDto'
        "404":
          description: No characteristics for this Gate available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateCharacteristicsDto'
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    EntityModelQubitDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelGateDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            gateDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelGateDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelGateDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        multiQubitGate:
          type: boolean
        operatingQubits:
          type: array
          items:
            type: string
            format: uuid
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelGateCharacteristicsDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            gateCharacteristicsDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelGateCharacteristicsDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelGateCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calibrationTime:
          type: string
          format: date-time
        gateTime:
          type: number
        gateErrorRate:
          type: number
        _links:
          $ref: '#/components/schemas/Links'
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean